feat: add KiMessageBox

+++ scripts/postinstall.mts
@@ -1,6 +1,7 @@
+import fs from 'node:fs'
 import spawn from 'cross-spawn'
 import { generateCode, loadFile, parseModule, writeFile } from 'magicast'
-import { cyan } from 'kolorist'
+import { cyan, red } from 'kolorist'
 import { deleteAsync } from 'del'
 import { isVue3 } from 'vue-demi'
 
@@ -13,17 +14,22 @@ async function postinstall() {
   }
 
   if (isVue3) {
-    console.log(cyan('Patching el-upload source code'))
-    const elUploadSourcePath = `${process.env.INIT_CWD}/node_modules/element-plus/es/components/upload/src/upload2.mjs`
-    let mod = await loadFile(elUploadSourcePath)
-    let { code } = generateCode(mod)
-    const whitespaces = code.match(/(?<=expose\({)\s*/)?.[0] || '\n'
-    code = code.replace(/expose\({(?!\s*uploadFiles,)/, `expose({${whitespaces}uploadFiles,`)
-    mod = parseModule(code)
-    await writeFile(mod, elUploadSourcePath)
+    const elementPlusPath = `${process.env.INIT_CWD}/node_modules/element-plus`
+    if (fs.existsSync(elementPlusPath)) {
+      console.log(red('[INFO] Patching el-upload source code'))
+      const elUploadSourcePath = `${elementPlusPath}/es/components/upload/src/upload2.mjs`
+      let mod = await loadFile(elUploadSourcePath)
+      let { code } = generateCode(mod)
+      const whitespaces = code.match(/(?<=expose\({)\s*/)?.[0] || '\n'
+      code = code.replace(/expose\({(?!\s*uploadFiles,)/, `expose({${whitespaces}uploadFiles,`)
+      mod = parseModule(code)
+      await writeFile(mod, elUploadSourcePath)
 
-    console.log(cyan('Re-bundling vite deps'))
-    console.log(await deleteAsync([`${process.env.INIT_CWD}/node_modules/.vite/deps/element-plus.js*`]))
+      console.log(cyan('[INFO] Re-bundling vite deps'))
+      await deleteAsync([`${process.env.INIT_CWD}/node_modules/.vite/deps/element-plus.js*`])
+    } else {
+      console.log(cyan('[INFO] You are using kikimore without element-plus'))
+    }
   }
 }
